+++ src/Header.js
@@ -1,0 +1,7 @@
+import "./Header.css";
+
+function Header() {
+    return <h2 className={"headerText"}>To Do List:</h2>
+}
+
+export default Header;
