+++ src/Tasks.css
@@ -1,0 +1,18 @@
+.tasks {
+    flex-grow: 1;
+}
+
+.react-tabs {
+    height: 100%;
+    background-color: #C1F4C5;
+    overflow: scroll;
+}
+
+.react-tabs__tab-list {
+    background-color: #65C18C;
+    position: sticky;
+}
+
+.react-tabs__tab.react-tabs__tab--selected {
+    background-color: #C1F4C5;
+}
