+++ src/index.css
@@ -1,13 +1,13 @@
 body {
-  margin: 0;
-  font-family: Futura, -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    margin: 0;
+    font-family: Futura, -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
 }
 
 code {
-  font-family: Futura, source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    font-family: Futura, source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
