--- src/oldIndex.css
@@ -1,210 +1,0 @@
-body {
-    font-family: Futura, sans-serif;
-    margin: 0px;
-    padding: 0px;
-}
-
-.app {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    height: 100vh;
-}
-
-.header {
-    position: sticky;
-    background-color: #65C18C;
-}
-
-.headerText {
-    font-size: 8vw;
-    margin: 2vw;
-    color: white;
-}
-
-.tasks {
-    flex-grow: 1;
-}
-
-.react-tabs {
-    height: 100%;
-    background-color: #C1F4C5;
-    overflow: scroll;
-}
-
-.react-tabs__tab-list {
-    background-color: #65C18C;
-    position: sticky;
-}
-
-.react-tabs__tab.react-tabs__tab--selected {
-    background-color: #C1F4C5;
-}
-
-.taskList {
-    display: flex;
-    flex-direction: column;
-    padding: 0.5vh 2vw;
-}
-
-.listItem {
-    background-color: white;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    column-gap: 0.2vw;
-    padding: 1vh;
-    margin: 0.5vh 0vw;
-    border-style: solid;
-    border-width: 1px;
-    border-radius: 5px;
-    word-wrap:anywhere;
-
-}
-
-.checkbox {
-}
-
-.person {
-    flex-grow: 10;
-}
-
-.editIcon {
-}
-
-.deleteIcon {
-}
-
-button.deleteCompleted{
-    font-family: Futura, sans-serif;
-    background-color: #FFD365;
-    width: 80%;
-    font-size: 12pt;
-    color: #6b2f02;
-    height: 5vh;
-    border-style: none;
-    border-radius: 20px 20px 20px 20px;
-    margin: 3vh auto;
-    display:block;
-}
-
-.footer {
-    position: sticky;
-    bottom: 0;
-    margin: auto;
-    padding-top: 20px;
-    padding-bottom: 20px;
-    width: 100%;
-    background-color: #65C18C;
-    z-index: 101;
-}
-
-.addTaskSection {
-    display: flex;
-    justify-content: center;
-}
-
-input.addTask {
-    font-family: Futura, sans-serif;
-    width: 70%;
-    border-width: 0px;
-    border-radius: 20px 0px 0px 20px;
-    border-color: #C1F4C5;
-    padding: 5px;
-}
-
-button.addTaskButton {
-    font-family: Futura, sans-serif;
-    background-color: #C1F4C5;
-    border-width: 0px;
-    border-radius: 0px 20px 20px 0px;
-    border-color: #C1F4C5;
-    width: 20%;
-    padding: 5px;
-}
-
-.person {
-    font-family: Futura, sans-serif;
-    padding: 0px;
-    margin: 0px;
-    border-top-width: 0px;
-    border-left-width: 0px;
-    border-right-width: 0px;
-    border-bottom-width: 0px;
-    /*border-style: solid;*/
-    /*border-color: black;*/
-    font-size: 12pt;
-    /*border-bottom: #65C18C;*/
-}
-
-.person.editing {
-    border-top-width: 0px;
-    border-left-width: 0px;
-    border-right-width: 0px;
-    border-bottom-width: 1px;
-    border-style: solid;
-    border-color: black;
-    font-size: 12pt;
-}
-
-.backdrop {
-    position: fixed;
-    top: 0;
-    left: 0;
-    bottom: 400pt;
-    right: 300pt;
-    width: 100vw;
-    height: 100vh;
-    background-color: rgba(0, 0, 0, .3);
-    z-index: 102;
-}
-
-.modal {
-    background-color: #fff;
-    border-radius: 5pt;
-    border-color: black;
-    border-style: solid;
-    margin: 0 auto;
-    top: 30vh;
-    height: 100pt;
-    width: 200pt;
-    padding: 10pt;
-    position: relative;
-    z-index: 103;
-}
-
-.alert-buttons {
-    position: absolute;
-    display: flex;
-    bottom: 0;
-    right: 0;
-    z-index: 104;
-}
-
-#completedTask{
-    text-decoration: line-through;
-    opacity: 1;
-    color: dimgray;
-}
-
-button.alert-button.alert-cancel {
-    font-family: Futura, sans-serif;
-    background-color: #dedede;
-    border-width: 0px;
-    border-radius: 20px 20px 20px 20px;
-    border-color: #C1F4C5;
-    padding: 5px 15px 5px 15px;
-    margin-bottom: 10px;
-    margin-right: 10px;
-}
-
-button.alert-button.alert-ok {
-    font-family: Futura, sans-serif;
-    background-color: #C1F4C5;
-    border-width: 0px;
-    border-radius: 20px 20px 20px 20px;
-    border-color: #C1F4C5;
-    padding: 5px 15px 5px 15px;
-    margin-bottom: 10px;
-    margin-right: 10px;
-}
